IoT/puller/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
google-cloud-pubsub==2.21.0
cryptography==42.0.5
